HARDWARE CHANGE - Complete change of ISR.
Ports D2 and D12 are now linked. D2 has been moved to A4. 
This is to allow D12 to trigger INT0 without needing to do a major rewiring. 

TIMER interrupts are disabled. Print routine and delays are now called by loop(). INT0 now handles parallel input.

This change creates the possibility that the ISR is called during the timing-sensitive send_code(...) routine. send_code can't be made atomic, as the parallel input is not buffered anywhere and the TPC65 does not use any kind of SYN/ACK for data but rather blind-fires to this controller. 

keycodes.h is a bit goofed and some characters, specifically '<' and '>' require the MOD key to be activated, which I don't have a routine for.